Refactor createClaimable operation in modules
