+++ nodes/Stellar/actions/entities/IPredicate.ts
@@ -2,7 +2,7 @@ export default interface IPredicate {
 	isPredicateConditional: boolean;
 	predicateType?: string;
 	isPredicateTimeRelative?: boolean;
-	predicateTimeValue?: string;
+	predicateTimeValue?: number;
 	predicate1?: { values: IPredicate };
 	predicate2?: { values: IPredicate };
 }
